I know where this is going I guess I'll go with it.
The issue is, I wanted some flyweight but its got to be opts and no more.

Given that context has to get passed, should I leave it implicit, or make it explicit? If I do make it explicit, how so? Well, this seems maybe ungood but I'm going with deferrant as the local context.